Move all update checks into main process
