+++ lib/electron-middleware.js
@@ -1,36 +1,12 @@
 /* globals window */
-const { shell, remote, ipcRenderer } = require( 'electron' );
-const semver = require( 'semver' );
-const { version } = require( '../package.json' );
-const { checkForUpdates } = require( '../lib/updates' );
+const { shell, ipcRenderer } = require( 'electron' );
 
 function openUrl( url ) {
 	shell.openExternal( url );
 }
 
-function updateCheck( dispatch ) {
-	checkForUpdates( { version, semver } )
-		.then( response => {
-			if ( ! response.updateAvailable ) {
-				remote.dialog.showMessageBox( {
-					type: 'info',
-					message: `You have the latest version of Gitnews, ${ response.oldVersion }!`,
-				} );
-				return;
-			}
-			const confirm = remote.dialog.showMessageBox( {
-				type: 'question',
-				message: 'A new version ' + response.newVersion + ' of Gitnews is available.',
-				detail: 'Do you want to download it now?',
-				buttons: [ 'Yes', 'No' ]
-			} );
-			if ( confirm === 0 ) {
-				dispatch( openUrl( response.newVersionUrl ) );
-			}
-		} )
-		.catch( err => {
-			console.error( 'Error while checking for updates:', err );
-		} );
+function updateCheck() {
+	ipcRenderer.send( 'check-for-updates' );
 }
 
 function setIcon( nextIcon ) {
