+++ lib/updates.js
@@ -1,4 +1,5 @@
 const getLatestRelease = require( 'get-latest-release' );
+const { get } = require( 'lodash' );
 
 function checkForNewVersion( { version, semver } ) {
 	return getLatestRelease( {
@@ -21,7 +22,7 @@ function getVersionComparator( { version, semver } ) {
 				newVersion,
 				oldVersion,
 				newVersionUrl: response.url,
-				newVersionDownloadUrl: response.download_url,
+				newVersionDownloadUrl: get( response, 'asset.download_url' ),
 			};
 		} catch ( err ) {
 			return Promise.reject( err );
@@ -29,6 +30,35 @@ function getVersionComparator( { version, semver } ) {
 	};
 }
 
+function checkForUpdates( { showCurrentVersion, version, semver, dialog, openUrl, download } ) {
+	checkForNewVersion( { version, semver } )
+		.then( response => {
+			if ( ! response.updateAvailable ) {
+				showCurrentVersion && dialog.showMessageBox( {
+					type: 'info',
+					message: `You have the latest version of Gitnews, ${ response.oldVersion }!`,
+				} );
+				return;
+			}
+			const confirm = dialog.showMessageBox( {
+				type: 'question',
+				message: 'A new version ' + response.newVersion + ' of Gitnews is available.',
+				detail: 'Do you want to download it now?',
+				buttons: [ 'Yes', 'No' ]
+			} );
+			if ( confirm === 0 ) {
+				if ( download && response.newVersionDownloadUrl ) {
+					download( response.newVersionDownloadUrl );
+				} else {
+					openUrl( response.newVersionUrl );
+				}
+			}
+		} )
+		.catch( err => {
+			console.error( 'Error while checking for updates:', err );
+		} );
+}
+
 module.exports = {
-	checkForUpdates: checkForNewVersion,
+	checkForUpdates,
 };
